Refer to remapped df
Hi there! 

First of all, thanks a lot for setting up this library in Python. We're eager to start using it! 

As I already saw you noted you're aware of the bugs and the potential issues with multiple markets, I still wanted to flag where it's breaking for us. I think you remap the columns in ip_df here, but don't reference the ip_df after. Therefore data is not yet in the expected format and you'd want to take ip_df. I'm not 100% sure though as the rest of the code is new to me, but I proposed changes for this issue so you could easily add it if you agree!

![git example](https://user-images.githubusercontent.com/37535349/122955293-46152e80-d380-11eb-9959-5e67f88e9533.png)

Thanks again and know your work is being used :)